renamed every occurrence from schoolClassSubject to schoolClassSubjects
