+++ frontend/src/app/exam/exam.component.html
@@ -1,5 +1,4 @@
-<oc-title-bar>
-    <span>{{ 'i18n.modules.exam.linkTitle' | translate }}</span>
+<oc-title-bar>occurrencecean>{{ 'i18n.modules.exam.linkTitle' | translate }}</span>
 </oc-title-bar>
 <ng-template #empty>
     <oc-empty hint="i18n.modules.exam.empty.hint"></oc-empty>
@@ -12,7 +11,7 @@
                 <mat-select formControlName="subjectId"
                             [placeholder]="'i18n.modules.exam.filter.subject' | translate"
                             floatPlaceholder="never">
-                    <mat-optgroup *ngFor="let schoolClass of schoolClassSubject"
+                    <mat-optgroup *ngFor="let schoolClass of schoolClassSubjects"
                                   [label]="schoolClass.schoolClass.normalizedName">
                         <mat-option *ngFor="let subject of schoolClass.subjects" [value]="subject.id">
                             {{ subject.name }}
