+++ frontend/src/app/exam/exam.component.spec.ts
@@ -30,7 +30,7 @@ describe('ExamComponent', () => {
                     provide: ActivatedRoute,
                     useValue: {
                         data: Observable.of({
-                            taskFilter: {schoolClassSubjects: []},
+                            schoolClassSubjects: [],
                             allExams: [],
                             activeExams: []
                         })
