+++ frontend/src/app/exam/service/active-exam-list-resolver.service.ts
