+++ frontend/src/app/exam/service/exam-list-resolver.service.ts
