+++ frontend/src/app/task/task.component.html
@@ -11,7 +11,7 @@
                 <mat-select formControlName="subjectId"
                             [placeholder]="'i18n.modules.exam.filter.subject' | translate"
                             floatPlaceholder="never">
-                    <mat-optgroup *ngFor="let schoolClass of schoolClassSubject"
+                    <mat-optgroup *ngFor="let schoolClass of schoolClassSubjects"
                                   [label]="schoolClass.schoolClass.normalizedName">
                         <mat-option *ngFor="let subject of schoolClass.subjects" [value]="subject.id">
                             {{ subject.name }}
